Updated visual tests
"yarn jest -u" fails for the test "expect(clipboardText).toBe(" on line 54 of "gallery\visualtests\nxCodeSnippet.js". Turns out this is a bug and created a ticket for it. I removed that test from the code to ensure "yarn jest -u" passes for all suites. This removal was not pushed to the PR. Any changes to nxCodeSnippet.js and its corresponding screenshots were discarded.